backend/requirements.txt
absl-py==1.4.0
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
annotated-types==0.5.0
anyio==3.7.1
appdirs==1.4.4
asgiref==3.8.1
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.2
bitsandbytes>0.37.0
bleach==6.1.0
blinker==1.6.2
build==1.2.1
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.2.0
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.4
clickhouse-connect==0.6.6
cohere==4.12.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.1.3
contourpy==1.1.0
cryptography==41.0.2
cycler==0.11.0
dataclasses-json==0.5.9
datasets==2.13.1
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.6
diskcache==5.6.3
dnspython==2.6.1
docker-pycreds==0.4.0
docopt==0.6.2
duckdb==0.8.1
einops==0.6.1
email_validator==2.1.1
esprima==4.0.1
et-xmlfile==1.1.0
exceptiongroup==1.1.2
executing==1.2.0
faiss-cpu==1.7.4
fastapi==0.110.1
fastjsonschema==2.19.1
ffmpy==0.3.0
filelock==3.12.2
filetype==1.2.0
flatbuffers==23.5.26
fonttools==4.40.0
frozenlist==1.3.3
fsspec==2023.6.0
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.22.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.63.0
gpt4all==1.0.3
greenlet==2.0.2
grpcio==1.62.1
h11==0.14.0
h5py==3.9.0
hf_transfer==0.1.4
hnswlib==0.7.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
huggingface-cli==0.1
huggingface-hub==0.20.1
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
importlib_resources==6.1.2
inquirerpy==0.3.4
InstructorEmbedding==1.0.1
ipykernel==6.24.0
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.1
Js2Py==0.74
jsonlines==3.1.0
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.18.0
jsonschema-specifications==2023.6.1
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyterlab_pygments==0.3.0
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
kubernetes==29.0.0
langchain==0.1.13
langchain-community==0.0.29
langchain-core==0.1.33
langchain-text-splitters==0.0.1
langchainhub==0.1.14
langsmith==0.1.31
libclang==16.0.0
linkify-it-py==2.0.2
llama_cpp_python==0.2.26
llvmlite==0.40.1
loguru==0.7.0
lxml==4.9.3
lz4==4.3.2
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
mistune==3.0.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.4
multiprocess==0.70.14
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
numba==0.57.1
numexpr==2.8.4
numpy==1.24.4
oauthlib==3.2.2
olefile==0.46
onnxruntime==1.15.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
opt-einsum==3.3.0
orjson==3.9.15
overrides==7.3.1
packaging==23.2
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.3
pathtools==0.1.2
pdf2image==1.16.3
pdfminer.six==20221105
pfzy==0.3.4
pickleshare==0.7.5
Pillow==9.5.0
pipreqs==0.5.0
pipwin==0.5.2
platformdirs==3.8.1
posthog==3.0.1
prettytable==3.8.0
prompt-toolkit==3.0.39
protobuf==4.23.4
psutil==5.9.5
pulsar-client==3.2.0
pure-eval==0.2.2
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.4
pydantic-extra-types==2.6.0
pydantic-settings==2.2.1
pydantic_core==2.16.3
pydeck==0.8.1b0
pydub==0.25.1
Pygments==2.15.1
pyjsparser==2.7.1
Pympler==1.0.1
PyMuPDF==1.22.5
pypandoc==1.11
pyparsing==3.0.9
pypdf==3.12.1
PyPika==0.48.9
PyPrind==2.11.3
pyproject_hooks==1.0.0
pyreadline3==3.4.1
pySmartDL==1.3.4
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-magic==0.4.27
python-multipart==0.0.9
python-pptx==0.6.21
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==25.1.0
referencing==0.29.1
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.4.2
rpds-py==0.8.10
rsa==4.9
ruff==0.3.0
safetensors==0.4.2
scikit-learn==1.3.0
scipy==1.11.1
seaborn==0.12.2
semantic-version==2.10.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
sentry-sdk==1.28.0
setproctitle==1.3.2
shellingham==1.5.4
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.18
stack-data==0.6.2
starlette==0.37.2
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.3.0
threadpoolctl==3.1.0
tiktoken==0.4.0
tinycss2==1.2.1
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers @ git+https://github.com/huggingface/transformers.git@831bc25d8fdb85768402f772cf65cc3d7872b211
tree-sitter==0.21.3
tree-sitter-languages==1.10.2
trl==0.4.6
typer==0.9.0
types-requests<=2.31.0.2
typing==3.7.4.3
typing-inspect==0.8.0
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==4.3.1
uc-micro-py==1.0.2
ujson==5.9.0
Unidecode==1.3.6
unstructured==0.8.1
urllib3==1.26.18
uvicorn==0.22.0
validators==0.20.0
wandb==0.15.5
watchdog==3.0.0
watchfiles==0.19.0
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.7.0
websockets==11.0.3
Werkzeug==2.3.6
win32-setctime==1.1.0
wonderwords==2.2.0
wrapt==1.15.0
xformers==0.0.21
xlrd==2.0.1
XlsxWriter==3.1.2
xxhash==3.2.0
yarg==0.1.9
yarl==1.9.2
zipp==3.16.0
zstandard==0.21.0
